fix(monitor): accept fractional monitor timeouts
The server stores `monitor.timeout` in a floating point column and the
web UI enters the value in steps of a tenth of a second, but every
timeout field was typed as an integer, so `encoding/json` refused a
value such as 0.5 and every typed conversion of the monitor failed.

Change `monitor.HTTPDetails.Timeout` and
`monitor.RealBrowserDetails.Timeout` to `float64` and
`monitor.KafkaProducerDetails.Timeout`,
`monitor.NTPDetails.Timeout`, `monitor.PingDetails.Timeout`,
`monitor.RabbitMQDetails.Timeout` and `monitor.SteamDetails.Timeout`
to `*float64`, together with the substituted defaults for the NOT NULL
column. Whole-second values still marshal without a fractional part.

This is a breaking change for consumers reading or setting a timeout.

+++ monitor/monitor_http.go
@@ -104,9 +104,13 @@ func (h HTTP) MarshalJSON() ([]byte, error) {
 }
 
 // HTTPDetails contains http-specific monitor configuration.
+//
+// Timeout is the request timeout in seconds. The server stores it in a
+// floating point column and the web UI enters it in steps of a tenth of a
+// second, so fractional values round-trip unchanged.
 type HTTPDetails struct {
 	URL                      string     `json:"url"`
-	Timeout                  int64      `json:"timeout"`
+	Timeout                  float64    `json:"timeout"`
 	ExpiryNotification       bool       `json:"expiryNotification"`
 	DomainExpiryNotification bool       `json:"domainExpiryNotification"`
 	IgnoreTLS                bool       `json:"ignoreTls"`
